fix color

+++ lib/widgets/editcolorwidget.dart
@@ -457,7 +457,7 @@ class _EditColorWidgetState extends State<EditColorWidget> {
       // 'price': _priceController.text,
       // 'product 3D model': model,
       'price': double.parse(_priceController.text),
-      'stocks': int.parse(_stocksController.text),
+      'stocks': double.parse(_stocksController.text),
     };
 
     // Add the product to Firestore
